RADICAL SIMPLIFICATION (check pls)

+++ src/utils/SphericalHarmonics.h
@@ -8,4 +8,4 @@ enum class Normalization
   SN3D,
 };
 
-std::array<double, 36> harmonics(const SphericalVector&, Normalization = Normalization::SN3D);
+std::array<float, 36> harmonics(const SphericalVector&, Normalization = Normalization::SN3D);
